First commit for GPT2 (#1)
Summary:
This a draft for a PR that implements (to the best of my knowledge) the first Opacus compatible version of GPT-2.

I am unable to get a reasonable privacy-perplexity trade-off when fine-tuning on wikitext103, but this may serve as a tough baseline for future DP-SGD improvements to aspire to surpass.

Pull Request resolved: https://github.com/facebookresearch/Opacus-lab/pull/1

Reviewed By: ashkan-software

Differential Revision: D30178860

Pulled By: ffuuugor

fbshipit-source-id: 8e857af86ed1885c5f93c9fbb79815c1658a165f